refactor(MetricWrapper): move to the evaluation subpackage
