+++ src/brisk/__init__.py
@@ -24,12 +24,12 @@ from brisk.defaults.classification_algorithms import CLASSIFICATION_ALGORITHMS
 from brisk.defaults.classification_metrics import CLASSIFICATION_METRICS
 from brisk.data.data_manager import DataManager
 from brisk.evaluation.metric_manager import MetricManager
+from brisk.evaluation.metric_wrapper import MetricWrapper
 from brisk.evaluation.evaluation_manager import EvaluationManager
 from brisk.reporting.report_manager import ReportManager
 from brisk.training.training_manager import TrainingManager
 from brisk.training.workflow import Workflow
 from brisk.utility.algorithm_wrapper import AlgorithmWrapper, AlgorithmCollection
-from brisk.utility.metric_wrapper import MetricWrapper
 from brisk.utility.alert_mailer import AlertMailer
 from brisk.utility.utility import create_metric
 from brisk.utility.logging_util import TqdmLoggingHandler, FileFormatter
